+++ Faker_company_name.v
@@ -1,5 +1,5 @@
 module faker
 
 pub fn (mut this Faker) company_name() string {
-    return this.random_line_in_file("company-name")
+    return this.random_line_in_file(file: "company-name")
 }