+++ Faker_company_name_test.v
@@ -5,5 +5,5 @@ import faker { Faker, get_rows_in_file }
 fn test_returns_random_company_name() {
     mut fake := Faker{}
 
-    assert get_rows_in_file("company-name", .en).contains(fake.company_name())
+    assert get_rows_in_file(file: "company-name", lang: .en).contains(fake.company_name())
 }