+++ Faker_first_name.v
@@ -1,5 +1,5 @@
 module faker
 
 pub fn (mut this Faker) first_name() string {
-    return this.random_line_in_file("first-name")
+    return this.random_line_in_file(file: "first-name")
 }