+++ Faker_last_name.v
@@ -1,5 +1,5 @@
 module faker
 
 pub fn (mut this Faker) last_name() string {
-    return this.random_line_in_file("last-name")
+    return this.random_line_in_file(file: "last-name")
 }