+++ Faker_sentence.v
@@ -1,5 +1,5 @@
 module faker
 
 pub fn (mut this Faker) sentence() string {
-    return this.random_line_in_file("sentence")
+    return this.random_line_in_file(file: "sentence")
 }