+++ Faker_using_lang_test.v
@@ -7,5 +7,5 @@ fn test_it_can_change_language_and_return_fake_data() {
 
     fake.using_lang(.en)
 
-    assert get_rows_in_file("sentence", .en).contains(fake.sentence())
+    assert get_rows_in_file(file: "sentence", lang: .en).contains(fake.sentence())
 }