+++ Faker_word.v
@@ -1,5 +1,5 @@
 module faker
 
 pub fn (mut this Faker) word() string {
-    return this.random_line_in_file("word")
+    return this.random_line_in_file(file: "word")
 }