+++ GetRowsInFileParameters.v
@@ -1,0 +1,8 @@
+module faker
+
+pub struct GetRowsInFileParameters {
+    pub:
+        file   string
+        not_tied_to_lang bool
+        lang Lang = .en
+}