+++ RandomLineInFileParameters.v
@@ -1,0 +1,6 @@
+module faker
+
+pub struct RandomLineInFileParameters {
+    file   string
+    not_tied_to_lang bool
+}